// src/core/models/show.ts
export interface Show {
  id: number;
  completedDate: string;
  createdAt: string;
  description: string;
  genre: string;
  imdbId: string;
  imdbRating: string;
  language: string;
  loved: boolean;
  showDate: string;
  startedDate: string;
  status: string;
  title: string;
  updatedAt: string;
  year: number;
}